Refactor creation of ACSAdapter
